refactor: standardizes file naming app
