+++ front-end-repository-searcher/src/app/features/repositories/pipes/capitalize.pipe.spec.ts
@@ -1,4 +1,4 @@
-import { CapitalizePipe } from './capitalize-pipe';
+import { CapitalizePipe } from './capitalize.pipe';
 
 describe('CapitalizePipe', () => {
   it('create an instance', () => {
