+++ front-end-repository-searcher/src/app/features/repositories/pipes/capitalize.pipe.ts
